+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added webpack for single file build output
 - Added change log
+- Added automatic unfeature of old daily posts
 
 ## 1.0.0 ##